extracted Jam Admin subcommands
