+++ bot/src/main/java/de/chojo/gamejam/commands/subcommands/SubCommand.java
@@ -1,0 +1,8 @@
+package de.chojo.gamejam.commands.subcommands;
+
+import de.chojo.jdautil.wrapper.SlashCommandContext;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+
+public interface SubCommand {
+    void execute(SlashCommandInteractionEvent event, SlashCommandContext context);
+}
